Update: Implementing conditionals and exceptions

+++ src/usecases/delete-user-account-usecase.ts
@@ -1,13 +1,15 @@
 import { IUserAccountRepository } from "../repositories/iUserAccountRepository";
+import { AppError } from "../infra/shared/errors/AppError";
 
 export class DeleteUserAccountUseCase {
   constructor(private userAccountRepository: IUserAccountRepository) {}
 
   async execute(id: string) {
-    // const checkUserID = await this.userAccountRepository.findById(id);
+    const checkUserID = await this.userAccountRepository.findById(id);
 
-    // if (!checkUserID) {
-    // }
+    if (checkUserID === null) {
+      throw new AppError("ID does not exist in the database!");
+    }
 
     await this.userAccountRepository.delete(id);
   }
